+++ core/auth/email.go
@@ -2,16 +2,33 @@ package auth
 
 import (
 	"core-building-block/core/model"
+	"core-building-block/utils"
+	"encoding/json"
+	"fmt"
+	"time"
 
 	"github.com/rokmetro/logging-library/errors"
 	"github.com/rokmetro/logging-library/logs"
 	"github.com/rokmetro/logging-library/logutils"
+	"golang.org/x/crypto/bcrypt"
 )
 
 const (
 	//AuthTypeEmail email auth type
 	AuthTypeEmail string = "email"
+
+	typeTime        logutils.MessageDataType = "time.Time"
+	typeEmailCreds  logutils.MessageDataType = "email creds"
+	typeEmailParams logutils.MessageDataType = "email params"
 )
+
+//enailCreds represents the creds struct for email auth
+type emailCreds struct {
+	Email              string    `json:"email" bson:"email" validate:"required"`
+	Password           string    `json:"password" bson:"password"`
+	VerificationCode   string    `json:"verification_code" bson:"verification_code"`
+	VerificationExpiry time.Time `json:"verification_expiry" bson:"verification_expiry"`
+}
 
 // Email implementation of authType
 type emailAuthImpl struct {
@@ -19,18 +36,159 @@ type emailAuthImpl struct {
 	authType string
 }
 
+func (a *emailAuthImpl) applySignUp(authType model.AuthType, appType model.ApplicationType, appOrg model.ApplicationOrganization, creds string, params string, l *logs.Log) (*string, map[string]interface{}, error) {
+	type signUpEmailParams struct {
+		Email           string `json:"email"`
+		Password        string `json:"password"`
+		ConfirmPassword string `json:"confirm_password"`
+	}
+	var sEmailParams signUpEmailParams
+	err := json.Unmarshal([]byte(params), &sEmailParams)
+	if err != nil {
+		return nil, nil, errors.WrapErrorAction("error getting sign_up email params", "", nil, err)
+	}
+	email := sEmailParams.Email
+	password := sEmailParams.Password
+	confirmPassword := sEmailParams.ConfirmPassword
+	if len(email) == 0 || len(password) == 0 || len(confirmPassword) == 0 {
+		return nil, nil, errors.WrapErrorAction("bad params data", "", nil, err)
+	}
+	//check if the passwrod matches with the confirm password one
+	if password != confirmPassword {
+		return nil, nil, errors.WrapErrorAction("passwords fields do not match", "", nil, err)
+	}
+
+	//TODO do not allow to register the user if already exists
+
+	//password hash
+	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
+	if err != nil {
+		return nil, nil, errors.WrapErrorAction(logutils.ActionCompute, model.TypeAuthCred, nil, errors.New("failed to generate hash from password"))
+	}
+
+	//verification code
+	code, err := utils.GenerateRandomString(64)
+	if err != nil {
+		return nil, nil, errors.WrapErrorAction(logutils.ActionCompute, model.TypeAuthCred, nil, errors.New("failed to generate random string for verify code"))
+
+	}
+
+	//send verification code
+	if err = a.sendVerificationCode(email, code, appOrg.Application.ID, appOrg.Organization.ID); err != nil {
+		return nil, nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAuthCred, nil, errors.New("failed to send verification email to user"))
+	}
+
+	emailCredValue := emailCreds{Email: email, Password: string(hashedPassword), VerificationCode: code, VerificationExpiry: time.Now().Add(time.Hour * 24)}
+	emailCredValueMap, err := emailCredsToMap(&emailCredValue)
+	if err != nil {
+		return nil, nil, errors.WrapErrorAction("failed email params to map", "", nil, err)
+	}
+
+	return &email, emailCredValueMap, nil
+}
+
+func (a *emailAuthImpl) checkCredentials(accountAuthType model.AccountAuthType, creds string, l *logs.Log) (*bool, error) {
+	//check is verified
+	if !accountAuthType.Credential.Verified {
+		return nil, errors.ErrorAction("not verified", "", nil)
+	}
+
+	//get stored credential
+	storedCreds, err := mapToEmailCreds(accountAuthType.Credential.Value)
+	if err != nil {
+		return nil, errors.WrapErrorAction("error on map to email creds", "", nil, err)
+	}
+
+	//get request credential
+	type signInPasswordCred struct {
+		Password string `json:"password"`
+	}
+	var sPasswordParams signInPasswordCred
+	err = json.Unmarshal([]byte(creds), &sPasswordParams)
+	if err != nil {
+		return nil, errors.WrapErrorAction("error getting sign_in password creds", "", nil, err)
+	}
+	requestPassword := sPasswordParams.Password
+
+	//compare stored and requets ones
+	err = bcrypt.CompareHashAndPassword([]byte(storedCreds.Password), []byte(requestPassword))
+	if err != nil {
+		return nil, errors.WrapErrorAction("bad credentials", "", nil, err)
+	}
+
+	valid := true
+	return &valid, nil
+}
+
+func (a *emailAuthImpl) sendVerificationCode(email string, verificationCode string, appID string, orgID string) error {
+	verificationLink := a.auth.host + fmt.Sprintf("/auth/verify/%v/%v/%v/%v/%v", a.authType, email, verificationCode, appID, orgID)
+	return a.auth.sendEmail(email, "Verify your email", "Please click the link below to verify your email:\n"+verificationLink, "")
+}
+
+//TODO: To be used in password reset flow
+// func (a *emailAuthImpl) sendPasswordReset(email string, password string) error {
+// 	return a.auth.SendEmail(email, "Password Reset", "Your temporary password is "+password, "")
+// }
+
+func (a *emailAuthImpl) verify(credential *model.Credential, verification string, l *logs.Log) (map[string]interface{}, error) {
+	credBytes, err := json.Marshal(credential.Value)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionMarshal, typeEmailCreds, nil, err)
+	}
+
+	var creds *emailCreds
+	err = json.Unmarshal(credBytes, &creds)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionUnmarshal, typeEmailCreds, nil, err)
+	}
+	err = a.compareVerifyCode(creds.VerificationCode, verification, creds.VerificationExpiry, l)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionValidate, model.TypeAuthCred, &logutils.FieldArgs{"verification_code": verification}, errors.New("invalid verification code"))
+	}
+
+	//Update verification data
+	creds.VerificationCode = ""
+	creds.VerificationExpiry = time.Time{}
+	credsMap, err := emailCredsToMap(creds)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionCast, typeEmailCreds, nil, err)
+	}
+
+	return credsMap, nil
+}
+
+func (a *emailAuthImpl) compareVerifyCode(credCode string, requestCode string, expiryTime time.Time, l *logs.Log) error {
+	if expiryTime.Before(time.Now()) {
+		return errors.WrapErrorAction(logutils.ActionValidate, typeTime, nil, errors.New("verification code has expired"))
+	}
+
+	if credCode != requestCode {
+		return errors.WrapErrorAction(logutils.ActionValidate, typeTime, nil, errors.New("Invalid verification code"))
+	}
+	return nil
+
+}
 func (a *emailAuthImpl) userExist(authType model.AuthType, appType model.ApplicationType, appOrg model.ApplicationOrganization, creds string, l *logs.Log) (*model.Account, *model.AccountAuthType, error) {
 	appID := appOrg.Application.ID
 	orgID := appOrg.Organization.ID
 	authTypeID := authType.ID
-	accountAuthTypeIdentifier := "silyana.y@inabyte.com" //TODO get it from the creds string
 
-	account, err := a.auth.storage.FindAccount(appID, orgID, authTypeID, accountAuthTypeIdentifier)
+	var requestCreds emailCreds
+	err := json.Unmarshal([]byte(creds), &requestCreds)
+	if err != nil {
+		return nil, nil, errors.WrapErrorAction(logutils.ActionUnmarshal, typeEmailCreds, logutils.StringArgs("request"), err)
+	}
+
+	account, err := a.auth.storage.FindAccount(appID, orgID, authTypeID, requestCreds.Email)
 	if err != nil {
 		return nil, nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAccount, nil, err) //TODO add args..
 	}
 
-	accountAuthType := account.FindAccountAuthType(authTypeID, accountAuthTypeIdentifier)
+	if account == nil {
+		return nil, nil, nil
+	}
+
+	accountAuthType, err := a.auth.findAccountAuthType(account, authTypeID, requestCreds.Email)
 	if accountAuthType == nil {
 		return nil, nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAccountAuthType, nil, err) //TODO add args..
 	}
@@ -38,11 +196,30 @@ func (a *emailAuthImpl) userExist(authType model.AuthType, appType model.Applica
 	return account, accountAuthType, nil
 }
 
-func (a *emailAuthImpl) checkCredentials(accountAuthType model.AccountAuthType, creds string, l *logs.Log) (*bool, error) {
-	//TODO - get the password from the creds and check it using user auth type id - from the credentials collection
+func emailCredsToMap(creds *emailCreds) (map[string]interface{}, error) {
+	credBytes, err := json.Marshal(creds)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionMarshal, model.TypeAuthCred, nil, err)
+	}
+	var credsMap map[string]interface{}
+	err = json.Unmarshal(credBytes, &credsMap)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionUnmarshal, model.TypeAuthCred, nil, err)
+	}
+	return credsMap, nil
+}
 
-	result := true
-	return &result, nil
+func mapToEmailCreds(credsMap map[string]interface{}) (*emailCreds, error) {
+	credBytes, err := json.Marshal(credsMap)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionMarshal, typeEmailCreds, nil, err)
+	}
+	var creds emailCreds
+	err = json.Unmarshal(credBytes, &creds)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionUnmarshal, typeEmailCreds, nil, err)
+	}
+	return &creds, nil
 }
 
 //initEmailAuth initializes and registers a new email auth instance