+++ core/auth/interfaces.go
@@ -8,12 +8,28 @@ import (
 	"github.com/rokmetro/auth-library/authorization"
 	"github.com/rokmetro/auth-library/tokenauth"
 	"github.com/rokmetro/logging-library/logs"
+	"go.mongodb.org/mongo-driver/mongo"
 )
 
 //authType is the interface for authentication for auth types which are not external for the system(the users do not come from external system)
 type authType interface {
+	//applySignUp applies sign up operation
+	// Returns:
+	//	identifier (*string): The unique identifier
+	//	credentialValue (map): Credential value
+	applySignUp(authType model.AuthType, appType model.ApplicationType, appOrg model.ApplicationOrganization, creds string, params string, l *logs.Log) (*string, map[string]interface{}, error)
+
+	//checks the verification code generated on email signup
+	// Returns:
+	//	authTypeCreds (map[string]interface{}): Updated Credential.Value
+	verify(credential *model.Credential, verification string, l *logs.Log) (map[string]interface{}, error)
+
 	//userExist checks if the user exists for application and organizations
+	// Returns:
+	//	account (*model.Account): User account
+	//	accountAuthType (*model.AccountAuthType): User account auth type
 	userExist(authType model.AuthType, appType model.ApplicationType, appOrg model.ApplicationOrganization, creds string, l *logs.Log) (*model.Account, *model.AccountAuthType, error)
+
 	//checkCredentials checks if the account credentials are valid for the account auth type
 	checkCredentials(accountAuthType model.AccountAuthType, creds string, l *logs.Log) (*bool, error)
 }
@@ -23,10 +39,8 @@ type authType interface {
 type externalAuthType interface {
 	//getLoginUrl retrieves and pre-formats a login url and params for the SSO provider
 	getLoginURL(authType model.AuthType, appType model.ApplicationType, appOrg model.ApplicationOrganization, redirectURI string, l *logs.Log) (string, map[string]interface{}, error)
-
 	//externalLogin logins in the external system and provides the authenticated user
 	externalLogin(authType model.AuthType, appType model.ApplicationType, appOrg model.ApplicationOrganization, creds string, params string, l *logs.Log) (*model.ExternalSystemUser, interface{}, error)
-
 	//userExist checks if the user exists
 	userExist(externalUserIdentifier string, authType model.AuthType, appType model.ApplicationType, appOrg model.ApplicationOrganization, l *logs.Log) (*model.Account, error)
 
@@ -62,7 +76,7 @@ type APIs interface {
 	//		Refresh Token (string): Refresh token that can be sent to refresh the access token once it expires
 	//		Account (Account): Account object for authenticated user
 	//		Params (interface{}): authType-specific set of parameters passed back to client
-	Login(authType string, creds string, appID string, orgID string, params string, l *logs.Log) (string, string, *model.Account, interface{}, error)
+	Login(authType string, creds string, appID string, orgID string, params string, l *logs.Log) (string, string, string, *model.Account, interface{}, error)
 
 	//Refresh refreshes an access token using a refresh token
 	//	Input:
@@ -73,6 +87,9 @@ type APIs interface {
 	//		Refresh token (string): Refresh token that can be sent to refresh the access token once it expires
 	//		Params (interface{}): authType-specific set of parameters passed back to client
 	Refresh(refreshToken string, l *logs.Log) (string, string, interface{}, error)
+
+	//Verify checks the verification code in the credentials collection
+	Verify(id string, verification string, l *logs.Log) error
 
 	//GetLoginURL returns a pre-formatted login url for SSO providers
 	//	Input:
@@ -151,8 +168,13 @@ type Storage interface {
 	FindOrganization(id string) (*model.Organization, error)
 
 	//Credentials
-	FindCredentialsByID(ID string) (*model.AuthCreds, error)
-	FindCredentials(orgID string, authType string, params map[string]interface{}) (*model.AuthCreds, error)
+	// FindCredentialsByID(ID string) (*model.AuthCreds, error)
+	// FindCredentials(orgID string, appID string, authType string, params map[string]interface{}) (*model.AuthCreds, error)
+	// UpdateCredentials(orgID string, appID string, authType string, creds *model.AuthCreds) error
+	// InsertCredentials(creds *model.AuthCreds, context mongo.SessionContext) error
+	FindCredential(ID string) (*model.Credential, error)
+	UpdateCredential(creds *model.Credential) error
+	InsertCredential(creds *model.Credential, context mongo.SessionContext) error
 
 	//RefreshTokens
 	FindRefreshToken(token string) (*model.AuthRefresh, error)