+++ core/model/auth.go
@@ -24,6 +24,8 @@ const (
 	TypeUserAuth logutils.MessageDataType = "user auth"
 	//TypeAuthCred auth cred type
 	TypeAuthCred logutils.MessageDataType = "auth cred"
+	//TypeCredential credential type
+	TypeCredential logutils.MessageDataType = "credential"
 	//TypeAuthRefresh auth refresh type
 	TypeAuthRefresh logutils.MessageDataType = "auth refresh"
 	//TypeRefreshToken refresh token type
@@ -105,6 +107,7 @@ type UserAuth struct {
 type AuthCreds struct {
 	ID        string                 `bson:"_id"`
 	OrgID     string                 `bson:"org_id"`
+	AppID     string                 `bson:"app_id"`
 	AuthType  string                 `bson:"auth_type"`
 	AccountID string                 `bson:"account_id"`
 	Creds     map[string]interface{} `bson:"creds"`